requirements.txt
absl-py
music21
xmltodict